+++ src/main/java/moonduck/server/repository/BoardSearchRepository.java
@@ -6,10 +6,13 @@ import com.querydsl.core.types.dsl.PathBuilder;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
-import moonduck.server.entity.QBoard;
 import moonduck.server.entity.Board;
+import moonduck.server.entity.Category;
+import moonduck.server.entity.QBoard;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.data.support.PageableExecutionUtils;
 import org.springframework.stereotype.Repository;
 
@@ -31,7 +34,6 @@ public class BoardSearchRepository {
                 .selectFrom(board)
                 .where(board.user.id.eq(userId))
                 .leftJoin(board.user).fetchJoin()
-                .leftJoin(board.program).fetchJoin()
                 .orderBy(getOrderSpecifier(filter))
                 .offset(pageable.getOffset())
                 .limit(pageable.getPageSize())
@@ -45,31 +47,15 @@ public class BoardSearchRepository {
         return PageableExecutionUtils.getPage(boards, pageable, countQuery::fetchOne);
     }
 
-    public Page<Board> findByUserIdAndCategoryWithFilter(Long userId, Board category, String filter, Pageable pageable) {
+    public List<Board> findByUserIdAndCategoryWithFilter(Long userId, Category category, String filter) {
         QBoard board = QBoard.board;
 
-        List<Board> boards = queryFactory
+        return queryFactory
                 .selectFrom(board)
-                .where(
-                        board.user.id.eq(userId),
-                        board.category.eq(category)
-                )
+                .where(board.user.id.eq(userId), board.category.eq(category))
                 .leftJoin(board.user).fetchJoin()
-                .leftJoin(board.program).fetchJoin()
                 .orderBy(getOrderSpecifier(filter))
-                .offset(pageable.getOffset())
-                .limit(pageable.getPageSize())
                 .fetch();
-
-        JPAQuery<Long> countQuery = queryFactory
-                .select(board.count())
-                .from(board)
-                .where(
-                        board.user.id.eq(userId),
-                        board.category.eq(category)
-                );
-
-        return PageableExecutionUtils.getPage(boards, pageable, countQuery::fetchOne);
     }
 
     private OrderSpecifier getOrderSpecifier(String filter) {
